# booking-service/api-docs.yaml
openapi: 3.0.0
info:
  title: Booking API
  description: API for managing bookings.
  version: 1.0.0
servers:
  - url: http://localhost:8080
paths:
  /bookings/create:
    post:
      summary: Create a new booking.
      requestBody:
        description: The details of the booking to be created.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBookingRequest'
            example:
              date: "2024-12-10"
              serviceId: "54218760-ae5d-45b9-9ceb-58d36d869021"
              userId: "54218760-ae5d-45b9-9ceb-58d36d869022"
              start: 540
              end: 600
      responses:
        '200':
          description: Booking successfully created.
        '400':
          description: Bad request due to invalid parameters or malformed input.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
              example:
                type: "about:blank"
                title: "Bad Request"
                status: 400
                detail: "Start time must be before end time."
        '500':
          description: Internal server error due to an unexpected condition.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'

  /bookings/delete:
    post:
      summary: Delete an existing booking.
      requestBody:
        description: The details of the booking to be deleted.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeleteBookingRequest'
            example:
              date: "2024-12-10"
              serviceId: "54218760-ae5d-45b9-9ceb-58d36d869021"
              bookingId: "54218760-ae5d-45b9-9ceb-58d36d869023"
              userId: "54218760-ae5d-45b9-9ceb-58d36d869022"
      responses:
        '200':
          description: Booking successfully deleted.
        '400':
          description: Bad request due to invalid parameters or malformed input.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
              example:
                type: "about:blank"
                title: "Bad Request"
                status: 400
                detail: "Booking ID does not exist."
        '500':
          description: Internal server error due to an unexpected condition.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'

  /bookings/available:
    get:
      summary: Get available time ranges for a service on a specific date.
      parameters:
        - name: serviceId
          in: query
          description: The ID of the service for which time ranges are being queried.
          required: true
          schema:
            type: string
            format: uuid
          example: "54218760-ae5d-45b9-9ceb-58d36d869021"
        - name: date
          in: query
          description: The date to check for available time ranges.
          required: true
          schema:
            type: string
            format: date
          example: "2024-12-10"
        - name: serviceDuration
          in: query
          description: The duration of the service in minutes.
          required: true
          schema:
            type: integer
            format: int32
            minimum: 1
          example: 30
      responses:
        '200':
          description: A list of available time ranges for the service.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TimeRangeResponse'
              example:
                - start: 540
                  end: 570
                - start: 600
                  end: 630
        '400':
          description: Bad request due to invalid parameters or malformed input.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
              examples:
                invalidUUID:
                  summary: Invalid UUID format
                  value:
                    type: "about:blank"
                    title: "Bad Request"
                    status: 400
                    detail: "Parameter: 'serviceId' has incorrect value: '54218760-ae5d-45b9-9ceb-58d36d86902X'"
                invalidDuration:
                  summary: Invalid service duration
                  value:
                    type: "about:blank"
                    title: "Bad Request"
                    status: 400
                    detail: "getAvailableTimeRanges.serviceDuration: must be greater than 0"
        '500':
          description: Internal server error due to an unexpected condition.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'

  /bookings/user:
    get:
      summary: Get the time range for a specific user booking.
      parameters:
        - name: serviceId
          in: query
          description: The ID of the service for the booking.
          required: true
          schema:
            type: string
            format: uuid
          example: "54218760-ae5d-45b9-9ceb-58d36d869021"
        - name: date
          in: query
          description: The date of the booking.
          required: true
          schema:
            type: string
            format: date
          example: "2024-12-10"
        - name: userId
          in: query
          description: The ID of the user who made the booking.
          required: true
          schema:
            type: string
            format: uuid
          example: "54218760-ae5d-45b9-9ceb-58d36d869022"
        - name: bookingId
          in: query
          description: The ID of the booking.
          required: true
          schema:
            type: string
            format: uuid
          example: "54218760-ae5d-45b9-9ceb-58d36d869023"
      responses:
        '200':
          description: User booking time range information.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimeRangeResponse'
              example:
                start: 540
                end: 600
        '400':
          description: Bad request due to invalid parameters or malformed input.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
              example:
                type: "about:blank"
                title: "Bad Request"
                status: 400
                detail: "Required request parameter 'userId' for method parameter type UUID is not present"
        '404':
          description: User booking not found.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
              example:
                type: "about:blank"
                title: "Not Found"
                status: 404
                detail: "UserBooking(date: 2024-12-10, serviceId: 54218760-ae5d-45b9-9ceb-58d36d869021, bookingId: 54218760-ae5d-45b9-9ceb-58d36d869023) was not found."
        '500':
          description: Internal server error due to an unexpected condition.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'

  /bookings:
    get:
      summary: Get user bookings with cursor-based pagination.
      description: Retrieve bookings for a specific user. Supports cursor-based pagination for efficient fetching of large result sets.
      parameters:
        - name: userId
          in: query
          description: The ID of the user whose bookings to retrieve.
          required: true
          schema:
            type: string
            format: uuid
          example: "54218760-ae5d-45b9-9ceb-58d36d869022"
        - name: cursorServiceId
          in: query
          description: The service ID cursor for pagination (required with cursorDate and cursorBookingId for next page).
          required: false
          schema:
            type: string
            format: uuid
          example: "54218760-ae5d-45b9-9ceb-58d36d869021"
        - name: cursorDate
          in: query
          description: The date cursor for pagination (required with cursorServiceId and cursorBookingId for next page).
          required: false
          schema:
            type: string
            format: date
          example: "2024-12-10"
        - name: cursorBookingId
          in: query
          description: The booking ID cursor for pagination (required with cursorServiceId and cursorDate for next page).
          required: false
          schema:
            type: string
            format: uuid
          example: "54218760-ae5d-45b9-9ceb-58d36d869023"
        - name: limit
          in: query
          description: Maximum number of bookings to return.
          required: true
          schema:
            type: integer
            format: int32
          example: 10
      responses:
        '200':
          description: A list of user bookings.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserBookingResponse'
              example:
                - serviceId: "54218760-ae5d-45b9-9ceb-58d36d869021"
                  date: "2024-12-01"
                  bookingId: "54218760-ae5d-45b9-9ceb-58d36d869023"
                  start: 540
                  end: 600
                - serviceId: "54218760-ae5d-45b9-9ceb-58d36d869021"
                  date: "2024-12-02"
                  bookingId: "54218760-ae5d-45b9-9ceb-58d36d869024"
                  start: 660
                  end: 720
        '400':
          description: Bad request due to invalid parameters or malformed input.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
              example:
                type: "about:blank"
                title: "Bad Request"
                status: 400
                detail: "Required request parameter 'userId' for method parameter type UUID is not present"
        '500':
          description: Internal server error due to an unexpected condition.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'

components:
  schemas:
    CreateBookingRequest:
      type: object
      properties:
        date:
          type: string
          format: date
          description: The date of the booking
          example: "2024-12-10"
        serviceId:
          type: string
          format: uuid
          description: The ID of the service to book
          example: "54218760-ae5d-45b9-9ceb-58d36d869021"
        userId:
          type: string
          format: uuid
          description: The ID of the user making the booking
          example: "54218760-ae5d-45b9-9ceb-58d36d869022"
        start:
          type: integer
          description: Start time in minutes from midnight (0-1439)
          example: 540
        end:
          type: integer
          description: End time in minutes from midnight (0-1439)
          example: 600
      required:
        - date
        - serviceId
        - userId
        - start
        - end

    DeleteBookingRequest:
      type: object
      properties:
        date:
          type: string
          format: date
          description: The date of the booking
          example: "2024-12-10"
        serviceId:
          type: string
          format: uuid
          description: The ID of the service
          example: "54218760-ae5d-45b9-9ceb-58d36d869021"
        bookingId:
          type: string
          format: uuid
          description: The ID of the booking to delete
          example: "54218760-ae5d-45b9-9ceb-58d36d869023"
        userId:
          type: string
          format: uuid
          description: The ID of the user who made the booking
          example: "54218760-ae5d-45b9-9ceb-58d36d869022"
      required:
        - date
        - serviceId
        - bookingId
        - userId

    TimeRangeResponse:
      type: object
      properties:
        start:
          type: integer
          description: Start time in minutes from midnight (0-1439)
          example: 540
        end:
          type: integer
          description: End time in minutes from midnight (0-1439)
          example: 600
      required:
        - start
        - end

    UserBookingResponse:
      type: object
      properties:
        serviceId:
          type: string
          format: uuid
          description: The ID of the booked service
          example: "54218760-ae5d-45b9-9ceb-58d36d869021"
        date:
          type: string
          format: date
          description: The date of the booking
          example: "2024-12-01"
        bookingId:
          type: string
          format: uuid
          description: The ID of the booking
          example: "54218760-ae5d-45b9-9ceb-58d36d869023"
        start:
          type: integer
          description: Start time in minutes from midnight (0-1439)
          example: 540
        end:
          type: integer
          description: End time in minutes from midnight (0-1439)
          example: 600
      required:
        - serviceId
        - date
        - bookingId
        - start
        - end

    ProblemDetail:
      type: object
      properties:
        type:
          type: string
          description: URI reference identifying the problem type
          example: "about:blank"
        title:
          type: string
          description: Short, human-readable summary of the problem
          example: "Bad Request"
        status:
          type: integer
          description: HTTP status code
          example: 400
        detail:
          type: string
          description: Human-readable explanation specific to this occurrence
          example: "Parameter: 'serviceId' has incorrect value: 'invalid-uuid'"
        instance:
          type: string
          description: URI reference identifying the specific occurrence of the problem
      required:
        - type
        - title
        - status
        - detail